feat: implement MongoDB connection and define NoteDB and VaultDB models
