+++ server/package.json
@@ -4,9 +4,8 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "start": "nodemon indexAWS.js",
-    "dev": "nodemon indexDev.js",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "start": "node index.js",
+    "dev": "nodemon app.dev.js"
   },
   "author": "",
   "license": "ISC",
@@ -18,6 +17,7 @@
     "jsonwebtoken": "^9.0.2",
     "mongoose": "^8.0.0",
     "nodemon": "^3.0.1",
+    "serverless-http": "^3.2.0",
     "validator": "^13.12.0"
   }
 }
