--- server/src/connectDB.js
@@ -1,13 +1,0 @@
-const mongoose = require('mongoose');
-
-const mongoConnect = async () => {
-    try {
-        await mongoose.connect(process.env.MONGO_URI);
-        console.log('MongoDB Connected Successfully!!');
-    }
-    catch (error) {
-        console.warn('MongoDB Connection FAILED!! \n', error);
-    }
-}
-
-module.exports = mongoConnect;
