--- server/src/model/UserDB.js
@@ -1,33 +1,0 @@
-const mongo = require("mongoose");
-
-const UserDB = new mongo.Schema({
-    email: {
-        type: String,
-        required: true,
-        unique: true,
-    },
-    password: {
-        type: String,
-        required: true,
-    },
-    name: {
-        type: String,
-    },
-    dateOfBirth: {
-        type: String,
-        required: true,
-    },
-    secretAnswer: {
-        type: String,
-        required: true,
-    },
-    textVerify: {
-        type: String,
-    },
-    createdAt: {
-        type: Date,
-        default: Date.now,
-    },
-});
-
-module.exports = mongo.model('UserDB', UserDB);
